Move tests to own class
